+++ ycc_hull/api/helpers.py
@@ -1,7 +1,9 @@
 """
 Helpers API endpoints.
 """
+
 from collections.abc import Sequence
+from datetime import date
 
 from fastapi import APIRouter, Depends
 from ycc_hull.api.errors import create_http_exception_403
@@ -18,25 +20,36 @@ api_helpers = APIRouter(dependencies=[Depends(auth)])
 controller = HelpersController()
 
 
-def _published(user: User) -> bool | None:
-    return None if user.helpers_app_admin_or_editor else True
-
-
 @api_helpers.get("/api/v1/helpers/task-categories")
 async def helper_task_categories_get() -> Sequence[HelperTaskCategoryDto]:
     return await controller.find_all_task_categories()
 
 
 @api_helpers.get("/api/v1/helpers/tasks")
-async def helper_tasks_get(user: User = Depends(auth)) -> Sequence[HelperTaskDto]:
-    return await controller.find_all_tasks(published=_published(user))
+async def helper_tasks_get(
+    year: int | None = None, user: User = Depends(auth)
+) -> Sequence[HelperTaskDto]:
+    if not _can_access_year(year, user):
+        error_message = (
+            f"You do not have permission to list tasks for {year}"
+            if year
+            else "You do not have permission to list all tasks"
+        )
+        raise create_http_exception_403(error_message)
+
+    return await controller.find_all_tasks(year=year, published=_published(user))
 
 
 @api_helpers.get("/api/v1/helpers/tasks/{task_id}")
 async def helper_tasks_get_by_id(
     task_id: int, user: User = Depends(auth)
 ) -> HelperTaskDto:
-    return await controller.get_task_by_id(task_id, published=_published(user))
+    task = await controller.get_task_by_id(task_id, published=_published(user))
+
+    if not _can_access_year(task.year, user):
+        raise create_http_exception_403("You do not have permission to view this task")
+
+    return task
 
 
 @api_helpers.post("/api/v1/helpers/tasks")
@@ -96,3 +109,19 @@ async def helper_tasks_sign_up_as_helper(
 ) -> HelperTaskDto:
     await controller.sign_up_as_helper(task_id, user)
     return await controller.get_task_by_id(task_id, published=True)
+
+
+def _published(user: User) -> bool | None:
+    return None if user.helpers_app_admin_or_editor else True
+
+
+def _can_access_year(year: int | None, user: User) -> bool:
+    current_year = date.today().year
+
+    return (
+        year == current_year
+        or user.admin
+        or user.committee_member
+        or user.helpers_app_admin
+        or user.helpers_app_editor
+    )