+++ ycc_hull/auth.py
@@ -1,6 +1,7 @@
 """
 Keycloak authentication components.
 """
+
 import logging
 
 from fastapi import Depends