+++ ycc_hull/db/context.py
@@ -1,6 +1,7 @@
 """
 Database context.
 """
+
 from typing import Any, Awaitable, TypeVar
 from collections.abc import Callable, Sequence
 
@@ -22,7 +23,7 @@ class DatabaseContext:
     Database context.
     """
 
-    def __init__(self, database_url: str, echo: bool | None = None) -> None:
+    def __init__(self, database_url: str, *, echo: bool | None = None) -> None:
         self._engine: AsyncEngine = create_async_engine(database_url, echo=echo)
         self.async_session = async_sessionmaker(self._engine)