+++ ycc_hull/models/helpers_dtos.py
@@ -1,6 +1,7 @@
 """
 Helpers API DTO classes.
 """
+
 from collections.abc import Sequence
 from datetime import datetime
 
@@ -63,6 +64,16 @@ class HelperTaskDto(CamelisedBaseModelWithEntity[HelperTaskEntity]):
     captain: "HelperTaskHelperDto | None"
     helpers: Sequence["HelperTaskHelperDto"]
 
+    @property
+    def year(self) -> int:
+        if self.starts_at:
+            return self.starts_at.year
+        if self.ends_at:
+            return self.ends_at.year
+        if self.deadline:
+            return self.deadline.year
+        raise ValueError("Missing timing")
+
     @classmethod
     async def create(cls, task: HelperTaskEntity) -> "HelperTaskDto":
         return await cls._create(task, await task.awaitable_attrs.long_description)
@@ -98,11 +109,11 @@ class HelperTaskDto(CamelisedBaseModelWithEntity[HelperTaskEntity]):
             ends_at=task.ends_at,
             deadline=task.deadline,
             urgent=task.urgent,
-            captain_required_licence_info=await LicenceInfoDto.create(
-                captain_required_licence_info
-            )
-            if captain_required_licence_info
-            else None,
+            captain_required_licence_info=(
+                await LicenceInfoDto.create(captain_required_licence_info)
+                if captain_required_licence_info
+                else None
+            ),
             helper_min_count=task.helper_min_count,
             helper_max_count=task.helper_max_count,
             published=task.published,
@@ -150,17 +161,23 @@ class HelperTaskMutationRequestDto(CamelisedBaseModel):
 
     @model_validator(mode="after")
     def check_timing(self) -> "HelperTaskMutationRequestDto":
-        valid_shift = (
-            self.starts_at
-            and self.ends_at
-            and not self.deadline
-            and self.starts_at < self.ends_at
-        )
-        valid_deadline = not self.starts_at and not self.ends_at and self.deadline
+        if self.starts_at and self.ends_at and not self.deadline:
+            # Shifts have extra conditions
+            if self.starts_at >= self.ends_at:
+                raise ValueError("Invalid timing: start time must be before end time")
+            if self.starts_at.year != self.ends_at.year:
+                raise ValueError(
+                    "Invalid timing: start and end time must be in the same year"
+                )
+        elif not self.starts_at and not self.ends_at and self.deadline:
+            # Nothing more to validate on one value
+            pass
+        else:
+            raise ValueError(
+                "Invalid timing: either specify both start and end time for a shift or a deadline for a task"
+            )
 
-        if valid_shift or valid_deadline:
-            return self
-        raise ValueError("Invalid timing")
+        return self
 
     @model_validator(mode="after")
     def check_helper_min_max_count(self) -> "HelperTaskMutationRequestDto":