+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Year-aware helper tasks
+
+### Changed
+
+- Better error reporting for task timing issues
+
 ## [1.0.0] - 2024-01-30
 
 ### Changed